# perf: bench hannoy vs hnswrs