--- a/benches/speed.rs
+++ b/benches/speed.rs
@@ -1,3 +1,8 @@
+use std::env;
+use std::fs::OpenOptions;
+use std::hint::black_box;
+use std::io::Write;
+
 use hannoy::Reader;
 use hannoy::{distances::Cosine, Database, Writer};
 use heed::{Env, EnvOpenOptions, RwTxn};
@@ -28,7 +33,6 @@ fn gen_vecs<const DIM: usize>(size: usize) -> Vec<[f32; DIM]> {
         .collect()
 }
 
-// hannoy
 fn setup_lmdb() -> Env {
     let temp_dir = tempdir().unwrap();
     let env = unsafe {
@@ -57,45 +61,76 @@ fn create_lmdb_and_fill_with_vecs<'a, const DIM: usize>(
     Ok((writer, wtxn, db))
 }
 
-// time hnsw search
 #[divan::bench(
-    consts = [768],
-    sample_count = 100,
-)]
-#[ignore = "reason"]
+        consts = [512, 768, 1536],
+    )]
 fn search_hannoy<const DIM: usize>(bencher: divan::Bencher) {
     // first build a vector db
     let env = setup_lmdb();
-    let vecs = gen_vecs::<DIM>(50000);
+    let vecs = gen_vecs::<DIM>(100_000);
     let (writer, mut wtxn, db) = create_lmdb_and_fill_with_vecs::<DIM>(&env, &vecs).unwrap();
     let mut rng = rng();
     let mut builder = writer.builder(&mut rng);
     builder.ef_construction(32).build::<M, M0>(&mut wtxn).unwrap();
     wtxn.commit().unwrap();
 
-    // reader should have a lifetime to this
     let rtxn = env.read_txn().unwrap();
+
+    // https://www.kernel.org/doc/Documentation/sysctl/vm.txt
+    fn drop_caches() -> std::io::Result<()> {
+        let mut f = OpenOptions::new().write(true).open("/proc/sys/vm/drop_caches")?;
+        f.write_all(b"3")?;
+        Ok(())
+    }
 
     bencher
         .with_inputs(|| {
+            drop_caches().unwrap();
             let mut query = [f32::default(); DIM];
             thread_rng().fill(&mut query);
-            // Reader::open can incur some system calls that mess with profiling
             let reader = Reader::<Cosine>::open(&rtxn, 0, db).unwrap();
             (reader, query)
         })
         .bench_local_values(|(reader, query)| {
-            reader.nns(10).ef_search(10).by_vector(&rtxn, &query).unwrap();
+            black_box(reader.nns(10).ef_search(10).by_vector(&rtxn, &query).unwrap());
+        });
+}
+
+#[divan::bench(
+        consts = [512, 768, 1536],
+    )]
+fn search_hannoy_in_cache<const DIM: usize>(bencher: divan::Bencher) {
+    // first build a vector db
+    let env = setup_lmdb();
+    let vecs = gen_vecs::<DIM>(100_000);
+    let (writer, mut wtxn, db) = create_lmdb_and_fill_with_vecs::<DIM>(&env, &vecs).unwrap();
+    let mut rng = rng();
+    let mut builder = writer.builder(&mut rng);
+    builder.ef_construction(32).build::<M, M0>(&mut wtxn).unwrap();
+    wtxn.commit().unwrap();
+
+    let rtxn = env.read_txn().unwrap();
+
+    // prefetch whole graph
+    env::set_var("HANNOY_READER_PREFETCH_MEMORY", format!("{}", 1024 * 1024 * 1024));
+    bencher
+        .with_inputs(|| {
+            let mut query = [f32::default(); DIM];
+            thread_rng().fill(&mut query);
+            let reader = Reader::<Cosine>::open(&rtxn, 0, db).unwrap();
+            (reader, query)
+        })
+        .bench_local_values(|(reader, query)| {
+            black_box(reader.nns(10).ef_search(10).by_vector(&rtxn, &query).unwrap());
         });
 }
 
 // hnsw-rs
 #[divan::bench(
-    consts = [768],
-    sample_count = 100,
+    consts = [512, 768, 1536],
 )]
 fn search_hnsw_rs<const DIM: usize>(bencher: divan::Bencher) {
-    let nb_elem = 50000;
+    let nb_elem = 100_000;
     let vecs = gen_vecs::<DIM>(nb_elem);
     let max_nb_connection = M0;
     let nb_layer = 16.min((nb_elem as f32).ln().trunc() as usize);
@@ -123,7 +158,7 @@ fn search_hnsw_rs<const DIM: usize>(bencher: divan::Bencher) {
         .bench_local_values(|query| {
             let knbn = 10;
             let ef_c = 10;
-            hnsw.search(&query, knbn, ef_c);
+            black_box(hnsw.search(&query, knbn, ef_c));
         });
 }