Provides benchmarks for comparing hannoy to [`hnsw_rs`](https://docs.rs/hnsw_rs/latest/hnsw_rs/)

```
Timer precision: 41 ns
speed                      fastest       │ slowest       │ median        │ mean          │ samples │ iters
├─ search_hannoy                         │               │               │               │         │
│  ├─ 512                  146.1 µs      │ 665 µs        │ 247 µs        │ 267.2 µs      │ 100     │ 100
│  ├─ 768                  174 µs        │ 1.135 ms      │ 320.6 µs      │ 356.2 µs      │ 100     │ 100
│  ╰─ 1536                 266 µs        │ 1.714 ms      │ 544.6 µs      │ 543.5 µs      │ 100     │ 100
├─ search_hannoy_in_cache                │               │               │               │         │
│  ├─ 512                  234.9 µs      │ 492.4 µs      │ 324.4 µs      │ 334.1 µs      │ 100     │ 100
│  ├─ 768                  210 µs        │ 534 µs        │ 292 µs        │ 305.4 µs      │ 100     │ 100
│  ╰─ 1536                 233.2 µs      │ 521.5 µs      │ 345.2 µs      │ 344.7 µs      │ 100     │ 100
╰─ search_hnsw_rs                        │               │               │               │         │
   ├─ 512                   setting number of points 50000
 setting number of points 100000
265.6 µs      │ 622 µs        │ 391.1 µs      │ 408.5 µs      │ 100     │ 100
   ├─ 768                   setting number of points 50000
 setting number of points 100000
386.3 µs      │ 844.1 µs      │ 532 µs        │ 552.6 µs      │ 100     │ 100
   ╰─ 1536                  setting number of points 50000
w setting number of points 100000
662.8 µs      │ 1.398 ms      │ 940.3 µs      │ 972 µs        │ 100     │ 100

```