@jianshu93 2025-10-08
how about for sift1m dataset, 128 dimensions? see hnsw_rs example here: https://github.com/jean-pierreBoth/hnswlib-rs/blob/master/examples/ann-sift1m-128-euclidean.rs

you need to modify line 52 to the path that you download sift1m-128-euclidean.hdf5:

wget http://ann-benchmarks.com/sift-128-euclidean.hdf5